feat: game page frontend

+++ backend/src/services/user.service.ts
@@ -2,17 +2,24 @@ import bcrypt from "bcrypt";
 import jwt from "jsonwebtoken";
 import errors from "@/errors/index";
 import userRepository from "@/repositories/user.repository";
-import { NewUser, NewLogin } from "@/protocols/user.protocol";
-import dotenv from "dotenv";
+import { NewUser, NewLogin, NewUserEntity } from "@/protocols/user.protocol";
 
-
-async function createUser({ name, email, image, password }: NewUser) {
+async function createUser({ name, email, image, password, confirmation }: NewUser) {
     const checkEmail = await userRepository.findByEmail(email);
     if (checkEmail) throw errors.duplicatedEmailError();
 
+    if (password !== confirmation) throw errors.invalidCredentialsError();
 
     const hashPassword: string = await bcrypt.hash(password, 10)
-    await userRepository.createUser({ name, email, image, password: hashPassword })
+
+    const data: NewUserEntity = {
+        name,
+        email,
+        image,
+        password: hashPassword
+    } 
+
+    await userRepository.createUser(data)
 }
 
 async function findUsers() {
@@ -33,9 +40,9 @@ async function loginUser({ email, password }: NewLogin) {
 
     const validPassword = await bcrypt.compare(password, user.password);
     if (!validPassword) throw errors.invalidCredentialsError();
-
     const userId = Number(user.id)
     const token = jwt.sign({ userId }, process.env.SECRET_KEY);
+
     const userInfo = await userRepository.loginUser(
         token,
         userId
@@ -44,9 +51,12 @@ async function loginUser({ email, password }: NewLogin) {
     return userInfo;
 }
 
-async function deleteUser(id: string) {
-    const userId = parseInt(id)
-    const user = await userRepository.findById(userId);
+async function deleteUser(id: number, userId: number) {
+
+    if (id !== userId) throw errors.unauthorizedError();
+
+    const user = await userRepository.findById(id);
+    console.log(user)
     if (!user) throw errors.notFoundError();
 
     await userRepository.deleteUser(userId);
